+++ tests/Exception/ValueObject/InvalidSizeDimensionsExceptionTest.php
@@ -6,11 +6,13 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace Meritoo\Test\Common\Exception\ValueObject;
 
+use Meritoo\Common\Enums\OopVisibility;
 use Meritoo\Common\Exception\ValueObject\InvalidSizeDimensionsException;
 use Meritoo\Common\Test\Base\BaseTestCase;
-use Meritoo\Common\Type\OopVisibilityType;
 
 /**
  * Test case of an exception used while dimensions of size, passed to the instance of Size class, are invalid
@@ -50,7 +52,7 @@ class InvalidSizeDimensionsExceptionTest extends BaseTestCase
     {
         static::assertConstructorVisibilityAndArguments(
             InvalidSizeDimensionsException::class,
-            OopVisibilityType::IS_PUBLIC,
+            OopVisibility::Public,
             3
         );
     }