+++ tests/ValueObject/BankAccountTest.php
@@ -6,10 +6,12 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace Meritoo\Test\Common\ValueObject;
 
+use Meritoo\Common\Enums\OopVisibility;
 use Meritoo\Common\Test\Base\BaseTestCase;
-use Meritoo\Common\Type\OopVisibilityType;
 use Meritoo\Common\ValueObject\BankAccount;
 
 /**
@@ -37,7 +39,7 @@ class BankAccountTest extends BaseTestCase
     {
         static::assertConstructorVisibilityAndArguments(
             BankAccount::class,
-            OopVisibilityType::IS_PUBLIC,
+            OopVisibility::Public,
             2,
             2
         );